requirements.txt
python-telegram-bot
requests
yt-dlp
instaloader
moviepy
python-dotenv
Flask==2.3.2